use 16 bit typed array for input parsing

+++ src/Types.ts
@@ -112,7 +112,7 @@ export interface ICompositionHelper {
 export interface IInputHandler {
   parse(data: string): void;
   parseBytes(data: Uint8Array): void;
-  print(data: Uint32Array, start: number, end: number): void;
+  print(data: Uint16Array, start: number, end: number): void;
 
   /** C0 BEL */ bell(): void;
   /** C0 LF */ lineFeed(): void;
@@ -465,7 +465,7 @@ export interface IParsingState {
 */
 export interface IDcsHandler {
   hook(collect: string, params: number[], flag: number): void;
-  put(data: Uint32Array, start: number, end: number): void;
+  put(data: Uint16Array, start: number, end: number): void;
   unhook(): void;
 }
 
@@ -482,9 +482,9 @@ export interface IEscapeSequenceParser extends IDisposable {
    * Parse string `data`.
    * @param data The data to parse.
    */
-  parse(data: Uint32Array, length: number): void;
+  parse(data: Uint16Array, length: number): void;
 
-  setPrintHandler(callback: (data: Uint32Array, start: number, end: number) => void): void;
+  setPrintHandler(callback: (data: Uint16Array, start: number, end: number) => void): void;
   clearPrintHandler(): void;
 
   setExecuteHandler(flag: string, callback: () => void): void;
